chore: delete redundant file at project
