+++ src/features/Discussion/components/DiscussionHub.tsx
@@ -26,7 +26,7 @@ import {
 } from "@mui/icons-material";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useAuth } from "../../../context/AuthContext";
-import { discussionService } from "../../../services/discussionService";
+import { discussionService } from "../services/discussionService";
 
 interface Discussion {
   id: string;
