+++ src/features/Discussion/services/discussionService.ts
@@ -8,6 +8,10 @@ export interface Answer {
     createdAt: string;
     updatedAt: string | null;
     isEdited: boolean;
+    upvotes: number;
+    downvotes: number;
+    score: number;
+    userVotes: { [userId: string]: 'upvote' | 'downvote' };
 }
 
 export interface Discussion {
@@ -19,6 +23,12 @@ export interface Discussion {
     createdAt: string;
     updatedAt: string | null;
     isEdited: boolean;
+    upvotes: number;
+    downvotes: number;
+    score: number;
+    userVotes: { [userId: string]: 'upvote' | 'downvote' };
+    views: number;
+    viewedBy: string[];
     answers: Answer[];
 }
 
@@ -45,20 +55,36 @@ export interface UpdateAnswerData {
     content: string;
 }
 
-export const discussionService = {
+export interface VoteData {
+    userId: string;
+    voteType: 'upvote' | 'downvote';
+}
 
+export const discussionService = {
+    // Get all discussions
     getAllDiscussions: async (): Promise<Discussion[]> => {
         const response = await api.get('/discussions');
-        return response.data;
+        return response.data.map((discussion: any) => ({
+            ...discussion,
+            views: discussion.views ?? 0,
+            viewedBy: discussion.viewedBy ?? []
+        }));
     },
 
-
+    // Get discussion by ID
     getDiscussionById: async (id: string): Promise<Discussion> => {
         const response = await api.get(`/discussions/${id}`);
-        return response.data;
+        const discussion = response.data;
+
+        // Ensure view tracking fields exist for legacy discussions
+        return {
+            ...discussion,
+            views: discussion.views ?? 0,
+            viewedBy: discussion.viewedBy ?? []
+        };
     },
 
-
+    // Generate next discussion ID
     generateNextDiscussionId: async (): Promise<string> => {
         const discussions = await discussionService.getAllDiscussions();
         const maxId = discussions.reduce((max, discussion) => {
@@ -68,7 +94,7 @@ export const discussionService = {
         return (maxId + 1).toString();
     },
 
-
+    // Generate next answer ID for a discussion
     generateNextAnswerId: async (discussionId: string): Promise<string> => {
         const discussion = await discussionService.getDiscussionById(discussionId);
         const maxId = discussion.answers.reduce((max, answer) => {
@@ -78,7 +104,7 @@ export const discussionService = {
         return (maxId + 1).toString();
     },
 
-
+    // Create new discussion
     createDiscussion: async (data: CreateDiscussionData): Promise<Discussion> => {
         const newId = await discussionService.generateNextDiscussionId();
 
@@ -88,6 +114,12 @@ export const discussionService = {
             createdAt: new Date().toISOString(),
             updatedAt: null,
             isEdited: false,
+            upvotes: 0,
+            downvotes: 0,
+            score: 0,
+            userVotes: {},
+            views: 0,
+            viewedBy: [],
             answers: []
         };
 
@@ -95,7 +127,28 @@ export const discussionService = {
         return response.data;
     },
 
+    // Track view for discussion
+    trackView: async (discussionId: string, userId: string): Promise<Discussion> => {
+        const discussion = await discussionService.getDiscussionById(discussionId);
 
+        // Don't track view if:
+        // 1. User is the author
+        // 2. User has already viewed this discussion
+        if (discussion.authorId === userId || discussion.viewedBy.includes(userId)) {
+            return discussion;
+        }
+
+        const updatedDiscussion = {
+            ...discussion,
+            views: discussion.views + 1,
+            viewedBy: [...discussion.viewedBy, userId]
+        };
+
+        const response = await api.put(`/discussions/${discussionId}`, updatedDiscussion);
+        return response.data;
+    },
+
+    // Update discussion
     updateDiscussion: async (id: string, data: UpdateDiscussionData): Promise<Discussion> => {
         const discussion = await discussionService.getDiscussionById(id);
 
@@ -111,12 +164,12 @@ export const discussionService = {
         return response.data;
     },
 
-
+    // Delete discussion
     deleteDiscussion: async (id: string): Promise<void> => {
         await api.delete(`/discussions/${id}`);
     },
 
-
+    // Add answer to discussion
     addAnswer: async (data: CreateAnswerData): Promise<Discussion> => {
         const discussion = await discussionService.getDiscussionById(data.discussionId);
         const newAnswerId = await discussionService.generateNextAnswerId(data.discussionId);
@@ -128,7 +181,11 @@ export const discussionService = {
             authorName: data.authorName,
             createdAt: new Date().toISOString(),
             updatedAt: null,
-            isEdited: false
+            isEdited: false,
+            upvotes: 0,
+            downvotes: 0,
+            score: 0,
+            userVotes: {}
         };
 
         const updatedDiscussion = {
@@ -140,7 +197,7 @@ export const discussionService = {
         return response.data;
     },
 
-
+    // Update answer
     updateAnswer: async (discussionId: string, answerId: string, data: UpdateAnswerData): Promise<Discussion> => {
         const discussion = await discussionService.getDiscussionById(discussionId);
 
@@ -164,11 +221,103 @@ export const discussionService = {
         return response.data;
     },
 
-
+    // Delete answer
     deleteAnswer: async (discussionId: string, answerId: string): Promise<Discussion> => {
         const discussion = await discussionService.getDiscussionById(discussionId);
 
         const updatedAnswers = discussion.answers.filter(answer => answer.id !== answerId);
+
+        const updatedDiscussion = {
+            ...discussion,
+            answers: updatedAnswers
+        };
+
+        const response = await api.put(`/discussions/${discussionId}`, updatedDiscussion);
+        return response.data;
+    },
+
+    // Vote on discussion
+    voteOnDiscussion: async (discussionId: string, data: VoteData): Promise<Discussion> => {
+        const discussion = await discussionService.getDiscussionById(discussionId);
+
+        const currentUserVote = discussion.userVotes[data.userId];
+        let newUpvotes = discussion.upvotes;
+        let newDownvotes = discussion.downvotes;
+        const newUserVotes = { ...discussion.userVotes };
+
+        // Remove previous vote if exists
+        if (currentUserVote === 'upvote') {
+            newUpvotes--;
+        } else if (currentUserVote === 'downvote') {
+            newDownvotes--;
+        }
+
+        // Apply new vote if different from current
+        if (currentUserVote !== data.voteType) {
+            if (data.voteType === 'upvote') {
+                newUpvotes++;
+            } else {
+                newDownvotes++;
+            }
+            newUserVotes[data.userId] = data.voteType;
+        } else {
+            // Remove vote if clicking same button
+            delete newUserVotes[data.userId];
+        }
+
+        const updatedDiscussion = {
+            ...discussion,
+            upvotes: newUpvotes,
+            downvotes: newDownvotes,
+            score: newUpvotes - newDownvotes,
+            userVotes: newUserVotes
+        };
+
+        const response = await api.put(`/discussions/${discussionId}`, updatedDiscussion);
+        return response.data;
+    },
+
+    // Vote on answer
+    voteOnAnswer: async (discussionId: string, answerId: string, data: VoteData): Promise<Discussion> => {
+        const discussion = await discussionService.getDiscussionById(discussionId);
+
+        const updatedAnswers = discussion.answers.map(answer => {
+            if (answer.id === answerId) {
+                const currentUserVote = answer.userVotes[data.userId];
+                let newUpvotes = answer.upvotes;
+                let newDownvotes = answer.downvotes;
+                const newUserVotes = { ...answer.userVotes };
+
+                // Remove previous vote if exists
+                if (currentUserVote === 'upvote') {
+                    newUpvotes--;
+                } else if (currentUserVote === 'downvote') {
+                    newDownvotes--;
+                }
+
+                // Apply new vote if different from current
+                if (currentUserVote !== data.voteType) {
+                    if (data.voteType === 'upvote') {
+                        newUpvotes++;
+                    } else {
+                        newDownvotes++;
+                    }
+                    newUserVotes[data.userId] = data.voteType;
+                } else {
+                    // Remove vote if clicking same button
+                    delete newUserVotes[data.userId];
+                }
+
+                return {
+                    ...answer,
+                    upvotes: newUpvotes,
+                    downvotes: newDownvotes,
+                    score: newUpvotes - newDownvotes,
+                    userVotes: newUserVotes
+                };
+            }
+            return answer;
+        });
 
         const updatedDiscussion = {
             ...discussion,
